Add git tags to build releases
This patch adds a new string to the boot-up message "Git version: xxx"
indicating the git revision number, if the Makefile was able to run 'git
describe'. First we see if we are on a tagged release, and use just the
version tag as version. If we're not an tagged release, we try to get
the exact hash. In case this fails, we try a fallback to get anything
from git describe.

If git describe is not available or the Makefile has trouble executing
it, we do not store the version at all.

Ideally we would overwrite the real version string, Marlin 1.0.0, but it
has to be investigated if that is worth it, as some 3rd parties may
relentlessly assume this string to identify Marlin.

Note, that 'unclean/dirty' builds are postfixed with the -dirty flag,
while this will never be seen in production, it is quite possible during
development. On the OLED display will wrap the 'irty' bit onto an
invisible line.

Signed-off-by: Olliver Schinagl <o.schinagl@ultimaker.com>